Revert protobuf warnings fix

+++ src/protobuf-c/protobuf-c.c
@@ -444,16 +444,16 @@ required_field_get_packed_size(const ProtobufCFieldDescriptor *field,
 	case PROTOBUF_C_TYPE_STRING: {
 		const char *str = *(char * const *) member;
 		size_t len = str ? strlen(str) : 0;
-		return rv + uint32_size((uint32_t) len) + len;
+		return rv + uint32_size(len) + len;
 	}
 	case PROTOBUF_C_TYPE_BYTES: {
 		size_t len = ((const ProtobufCBinaryData *) member)->len;
-		return rv + uint32_size((uint32_t) len) + len;
+		return rv + uint32_size(len) + len;
 	}
 	case PROTOBUF_C_TYPE_MESSAGE: {
 		const ProtobufCMessage *msg = *(ProtobufCMessage * const *) member;
 		size_t subrv = msg ? protobuf_c_message_get_packed_size(msg) : 0;
-		return rv + uint32_size((uint32_t) subrv) + subrv;
+		return rv + uint32_size(subrv) + subrv;
 	}
 	}
 	PROTOBUF_C__ASSERT_NOT_REACHED();
@@ -593,26 +593,26 @@ repeated_field_get_packed_size(const ProtobufCFieldDescriptor *field,
 	case PROTOBUF_C_TYPE_STRING:
 		for (i = 0; i < count; i++) {
 			size_t len = strlen(((char **) array)[i]);
-			rv += uint32_size((uint32_t) len) + len;
+			rv += uint32_size(len) + len;
 		}
 		break;
 	case PROTOBUF_C_TYPE_BYTES:
 		for (i = 0; i < count; i++) {
 			size_t len = ((ProtobufCBinaryData *) array)[i].len;
-			rv += uint32_size((uint32_t) len) + len;
+			rv += uint32_size(len) + len;
 		}
 		break;
 	case PROTOBUF_C_TYPE_MESSAGE:
 		for (i = 0; i < count; i++) {
 			size_t len = protobuf_c_message_get_packed_size(
 				((ProtobufCMessage **) array)[i]);
-			rv += uint32_size((uint32_t) len) + len;
+			rv += uint32_size(len) + len;
 		}
 		break;
 	}
 
 	if (0 != (field->flags & PROTOBUF_C_FIELD_FLAG_PACKED))
-		header_size += uint32_size((uint32_t) rv);
+		header_size += uint32_size(rv);
 	return header_size + rv;
 }
 
@@ -916,7 +916,7 @@ string_pack(const char *str, uint8_t *out)
 		return 1;
 	} else {
 		size_t len = strlen(str);
-		size_t rv = uint32_pack((uint32_t) len, out);
+		size_t rv = uint32_pack(len, out);
 		memcpy(out + rv, str, len);
 		return rv + len;
 	}
@@ -937,7 +937,7 @@ static inline size_t
 binary_data_pack(const ProtobufCBinaryData *bd, uint8_t *out)
 {
 	size_t len = bd->len;
-	size_t rv = uint32_pack((uint32_t) len, out);
+	size_t rv = uint32_pack(len, out);
 	memcpy(out + rv, bd->data, len);
 	return rv + len;
 }
@@ -961,10 +961,10 @@ prefixed_message_pack(const ProtobufCMessage *message, uint8_t *out)
 		return 1;
 	} else {
 		size_t rv = protobuf_c_message_pack(message, out + 1);
-		uint32_t rv_packed_size = (uint32_t) uint32_size((uint32_t) rv);
+		uint32_t rv_packed_size = uint32_size(rv);
 		if (rv_packed_size != 1)
 			memmove(out + rv_packed_size, out + 1, rv);
-		return uint32_pack((uint32_t) rv, out) + rv;
+		return uint32_pack(rv, out) + rv;
 	}
 }
 
@@ -1267,11 +1267,11 @@ repeated_field_pack(const ProtobufCFieldDescriptor *field,
 
 		if (count == 0)
 			return 0;
-		header_len = (unsigned) tag_pack(field->id, out);
+		header_len = tag_pack(field->id, out);
 		out[0] |= PROTOBUF_C_WIRE_TYPE_LENGTH_PREFIXED;
 		len_start = header_len;
-		min_length = (unsigned) (get_type_min_size(field->type) * count);
-		length_size_min = (unsigned) uint32_size(min_length);
+		min_length = get_type_min_size(field->type) * count;
+		length_size_min = uint32_size(min_length);
 		header_len += length_size_min;
 		payload_at = out + header_len;
 
@@ -1279,13 +1279,13 @@ repeated_field_pack(const ProtobufCFieldDescriptor *field,
 		case PROTOBUF_C_TYPE_SFIXED32:
 		case PROTOBUF_C_TYPE_FIXED32:
 		case PROTOBUF_C_TYPE_FLOAT:
-			copy_to_little_endian_32(payload_at, array, (unsigned) count);
+			copy_to_little_endian_32(payload_at, array, count);
 			payload_at += count * 4;
 			break;
 		case PROTOBUF_C_TYPE_SFIXED64:
 		case PROTOBUF_C_TYPE_FIXED64:
 		case PROTOBUF_C_TYPE_DOUBLE:
-			copy_to_little_endian_64(payload_at, array, (unsigned) count);
+			copy_to_little_endian_64(payload_at, array, count);
 			payload_at += count * 8;
 			break;
 		case PROTOBUF_C_TYPE_INT32: {
@@ -1330,8 +1330,8 @@ repeated_field_pack(const ProtobufCFieldDescriptor *field,
 			PROTOBUF_C__ASSERT_NOT_REACHED();
 		}
 
-		payload_len = (unsigned) (payload_at - (out + header_len));
-		actual_length_size = (unsigned) uint32_size(payload_len);
+		payload_len = payload_at - (out + header_len);
+		actual_length_size = uint32_size(payload_len);
 		if (length_size_min != actual_length_size) {
 			assert(actual_length_size == length_size_min + 1);
 			memmove(out + header_len + 1, out + header_len,
@@ -1344,7 +1344,7 @@ repeated_field_pack(const ProtobufCFieldDescriptor *field,
 		/* not "packed" cased */
 		/* CONSIDER: optimize this case a bit (by putting the loop inside the switch) */
 		size_t rv = 0;
-		unsigned siz = (unsigned) sizeof_elt_in_repeated_array(field->type);
+		unsigned siz = sizeof_elt_in_repeated_array(field->type);
 
 		for (i = 0; i < count; i++) {
 			rv += required_field_pack(field, array, out + rv);
@@ -1487,7 +1487,7 @@ required_field_pack_to_buffer(const ProtobufCFieldDescriptor *field,
 		size_t sublen = str ? strlen(str) : 0;
 
 		scratch[0] |= PROTOBUF_C_WIRE_TYPE_LENGTH_PREFIXED;
-		rv += uint32_pack((uint32_t) sublen, scratch + rv);
+		rv += uint32_pack(sublen, scratch + rv);
 		buffer->append(buffer, rv, scratch);
 		buffer->append(buffer, sublen, (const uint8_t *) str);
 		rv += sublen;
@@ -1498,7 +1498,7 @@ required_field_pack_to_buffer(const ProtobufCFieldDescriptor *field,
 		size_t sublen = bd->len;
 
 		scratch[0] |= PROTOBUF_C_WIRE_TYPE_LENGTH_PREFIXED;
-		rv += uint32_pack((uint32_t) sublen, scratch + rv);
+		rv += uint32_pack(sublen, scratch + rv);
 		buffer->append(buffer, rv, scratch);
 		buffer->append(buffer, sublen, bd->data);
 		rv += sublen;
@@ -1516,7 +1516,7 @@ required_field_pack_to_buffer(const ProtobufCFieldDescriptor *field,
 			sublen = 0;
 		else
 			sublen = protobuf_c_message_pack_to_buffer(msg, &simple_buffer.base);
-		rv += uint32_pack((uint32_t) sublen, scratch + rv);
+		rv += uint32_pack(sublen, scratch + rv);
 		buffer->append(buffer, rv, scratch);
 		buffer->append(buffer, sublen, simple_buffer.data);
 		rv += sublen;
@@ -1716,14 +1716,14 @@ pack_buffer_packed_payload(const ProtobufCFieldDescriptor *field,
 #endif
 	case PROTOBUF_C_TYPE_INT32:
 		for (i = 0; i < count; i++) {
-			unsigned len = (unsigned) int32_pack(((int32_t *) array)[i], scratch);
+			unsigned len = int32_pack(((int32_t *) array)[i], scratch);
 			buffer->append(buffer, len, scratch);
 			rv += len;
 		}
 		break;
 	case PROTOBUF_C_TYPE_SINT32:
 		for (i = 0; i < count; i++) {
-			unsigned len = (unsigned) sint32_pack(((int32_t *) array)[i], scratch);
+			unsigned len = sint32_pack(((int32_t *) array)[i], scratch);
 			buffer->append(buffer, len, scratch);
 			rv += len;
 		}
@@ -1731,14 +1731,14 @@ pack_buffer_packed_payload(const ProtobufCFieldDescriptor *field,
 	case PROTOBUF_C_TYPE_ENUM:
 	case PROTOBUF_C_TYPE_UINT32:
 		for (i = 0; i < count; i++) {
-			unsigned len = (unsigned) uint32_pack(((uint32_t *) array)[i], scratch);
+			unsigned len = uint32_pack(((uint32_t *) array)[i], scratch);
 			buffer->append(buffer, len, scratch);
 			rv += len;
 		}
 		break;
 	case PROTOBUF_C_TYPE_SINT64:
 		for (i = 0; i < count; i++) {
-			unsigned len = (unsigned) sint64_pack(((int64_t *) array)[i], scratch);
+			unsigned len = sint64_pack(((int64_t *) array)[i], scratch);
 			buffer->append(buffer, len, scratch);
 			rv += len;
 		}
@@ -1746,14 +1746,14 @@ pack_buffer_packed_payload(const ProtobufCFieldDescriptor *field,
 	case PROTOBUF_C_TYPE_INT64:
 	case PROTOBUF_C_TYPE_UINT64:
 		for (i = 0; i < count; i++) {
-			unsigned len = (unsigned) uint64_pack(((uint64_t *) array)[i], scratch);
+			unsigned len = uint64_pack(((uint64_t *) array)[i], scratch);
 			buffer->append(buffer, len, scratch);
 			rv += len;
 		}
 		break;
 	case PROTOBUF_C_TYPE_BOOL:
 		for (i = 0; i < count; i++) {
-			unsigned len = (unsigned) boolean_pack(((protobuf_c_boolean *) array)[i], scratch);
+			unsigned len = boolean_pack(((protobuf_c_boolean *) array)[i], scratch);
 			buffer->append(buffer, len, scratch);
 			rv += len;
 		}
@@ -1784,7 +1784,7 @@ repeated_field_pack_to_buffer(const ProtobufCFieldDescriptor *field,
 		size_t tmp;
 
 		scratch[0] |= PROTOBUF_C_WIRE_TYPE_LENGTH_PREFIXED;
-		rv += uint32_pack((uint32_t) payload_len, scratch + rv);
+		rv += uint32_pack(payload_len, scratch + rv);
 		buffer->append(buffer, rv, scratch);
 		tmp = pack_buffer_packed_payload(field, count, array, buffer);
 		assert(tmp == payload_len);
@@ -1856,7 +1856,7 @@ protobuf_c_message_pack_to_buffer(const ProtobufCMessage *message,
 		} else {
 			rv += repeated_field_pack_to_buffer(
 				field,
-				(unsigned) (*(const size_t *) qmember),
+				*(const size_t *) qmember,
 				member,
 				buffer
 			);
@@ -1924,7 +1924,7 @@ parse_tag_and_wiretype(size_t len,
 		       uint32_t *tag_out,
 		       ProtobufCWireType *wiretype_out)
 {
-	unsigned max_rv = len > 5 ? 5 : (unsigned) len;
+	unsigned max_rv = len > 5 ? 5 : len;
 	uint32_t tag = (data[0] & 0x7f) >> 3;
 	unsigned shift = 4;
 	unsigned rv;
@@ -1964,7 +1964,7 @@ static inline uint32_t
 scan_length_prefixed_data(size_t len, const uint8_t *data,
 			  size_t *prefix_len_out)
 {
-	unsigned hdr_max = len < 5 ? (unsigned) len : 5;
+	unsigned hdr_max = len < 5 ? len : 5;
 	unsigned hdr_len;
 	uint32_t val = 0;
 	unsigned i;
@@ -2347,7 +2347,7 @@ parse_required_member(ScannedMember *scanned_member,
 		      ProtobufCAllocator *allocator,
 		      protobuf_c_boolean maybe_clear)
 {
-	unsigned len = (unsigned) scanned_member->len;
+	unsigned len = scanned_member->len;
 	const uint8_t *data = scanned_member->data;
 	ProtobufCWireType wire_type = scanned_member->wire_type;
 
@@ -2631,7 +2631,7 @@ parse_packed_repeated_member(ScannedMember *scanned_member,
 #endif
 	case PROTOBUF_C_TYPE_INT32:
 		while (rem > 0) {
-			unsigned s = scan_varint((unsigned) rem, at);
+			unsigned s = scan_varint(rem, at);
 			if (s == 0) {
 				PROTOBUF_C_UNPACK_ERROR("bad packed-repeated int32 value");
 				return FALSE;
@@ -2643,7 +2643,7 @@ parse_packed_repeated_member(ScannedMember *scanned_member,
 		break;
 	case PROTOBUF_C_TYPE_SINT32:
 		while (rem > 0) {
-			unsigned s = scan_varint((unsigned) rem, at);
+			unsigned s = scan_varint(rem, at);
 			if (s == 0) {
 				PROTOBUF_C_UNPACK_ERROR("bad packed-repeated sint32 value");
 				return FALSE;
@@ -2656,7 +2656,7 @@ parse_packed_repeated_member(ScannedMember *scanned_member,
 	case PROTOBUF_C_TYPE_ENUM:
 	case PROTOBUF_C_TYPE_UINT32:
 		while (rem > 0) {
-			unsigned s = scan_varint((unsigned) rem, at);
+			unsigned s = scan_varint(rem, at);
 			if (s == 0) {
 				PROTOBUF_C_UNPACK_ERROR("bad packed-repeated enum or uint32 value");
 				return FALSE;
@@ -2669,7 +2669,7 @@ parse_packed_repeated_member(ScannedMember *scanned_member,
 
 	case PROTOBUF_C_TYPE_SINT64:
 		while (rem > 0) {
-			unsigned s = scan_varint((unsigned) rem, at);
+			unsigned s = scan_varint(rem, at);
 			if (s == 0) {
 				PROTOBUF_C_UNPACK_ERROR("bad packed-repeated sint64 value");
 				return FALSE;
@@ -2682,7 +2682,7 @@ parse_packed_repeated_member(ScannedMember *scanned_member,
 	case PROTOBUF_C_TYPE_INT64:
 	case PROTOBUF_C_TYPE_UINT64:
 		while (rem > 0) {
-			unsigned s = scan_varint((unsigned) rem, at);
+			unsigned s = scan_varint(rem, at);
 			if (s == 0) {
 				PROTOBUF_C_UNPACK_ERROR("bad packed-repeated int64/uint64 value");
 				return FALSE;
@@ -2977,7 +2977,7 @@ protobuf_c_message_unpack(const ProtobufCMessageDescriptor *desc,
 
 		switch (wire_type) {
 		case PROTOBUF_C_WIRE_TYPE_VARINT: {
-			unsigned max_len = rem < 10 ? (unsigned) rem : 10;
+			unsigned max_len = rem < 10 ? rem : 10;
 			unsigned i;
 
 			for (i = 0; i < max_len; i++)
@@ -3082,7 +3082,7 @@ protobuf_c_message_unpack(const ProtobufCMessageDescriptor *desc,
 					      field->quantifier_offset);
 			if (*n_ptr != 0) {
                 void *a = 0;
-				unsigned n = (unsigned) *n_ptr;
+				unsigned n = *n_ptr;
 				*n_ptr = 0;
 				assert(rv->descriptor != NULL);
 #define CLEAR_REMAINING_N_PTRS()                                              \
